requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
PyQt5>=5.15.0
PyQt5-Qt5>=5.15.0
PyQt5-sip>=12.12.0
tensorboard>=2.14.0
pycocotools>=2.0.6
tqdm>=4.66.0
PyYAML>=6.0
requests>=2.31.0
albumentations>=1.3.0
